refactor: TrackedTsunamiTimeline の変換ロジックを logic 層へ分離
