+++ app/lib/feature/tsunami/data/notifier/tsunami_telegram_timeline_notifier.dart
@@ -1,6 +1,7 @@
 import 'package:eqmonitor/core/api/api_client_provider.dart';
+import 'package:eqmonitor/feature/tsunami/data/logic/tracked_tsunami_timeline_to_public_mapper.dart';
+import 'package:eqmonitor/feature/tsunami/data/logic/tsunami_telegrams_response_to_tracked_timeline_builder.dart';
 import 'package:eqmonitor/feature/tsunami/data/model/timeline/tsunami_timeline.dart';
-import 'package:eqmonitor/feature/tsunami/data/model/tracking/tracked_tsunami_timeline.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'tsunami_telegram_timeline_notifier.g.dart';
