+++ app/test/feature/tsunami/tsunami_api_to_domain_test.dart
@@ -1,4 +1,4 @@
-import 'package:eqmonitor/feature/tsunami/data/model/tracking/tracked_tsunami_timeline.dart';
+import 'package:eqmonitor/feature/tsunami/data/logic/tsunami_telegrams_response_to_tracked_timeline_builder.dart';
 import 'package:eqmonitor/feature/tsunami/data/model/tsunami_estimation_first_height.dart';
 import 'package:eqmonitor/feature/tsunami/data/model/tsunami_estimation_max_height.dart';
 import 'package:eqmonitor/feature/tsunami/data/model/tsunami_forecast_first_height.dart';
